code cleanup; merging chain statistical analysis form into sampling design form
