+++ webapp/views/App/views/Analysis/Chain/Chain.js
@@ -21,7 +21,6 @@ import ButtonBar from './ButtonBar'
 import { AnalysisNodeDefs } from './AnalysisNodeDefs'
 import { ChainBasicProps } from './ChainBasicProps'
 import { ChainSamplingDesignProps } from './ChainSamplingDesignProps'
-import { ChainStatisticalAnalysisProps } from './statisticalAnalysis/ChainStatisticalAnalysisProps'
 
 const ChainComponent = () => {
   const dispatch = useDispatch()
@@ -79,11 +78,6 @@ const ChainComponent = () => {
           {
             label: 'chainView.samplingDesign',
             component: ChainSamplingDesignProps,
-            props: { updateChain },
-          },
-          {
-            label: 'chainView.statisticalAnalysis.header',
-            component: ChainStatisticalAnalysisProps,
             props: { updateChain },
           },
         ]}
