+++ webapp/views/App/views/Analysis/Chain/DimensionsSelector/DimensionsSelector.js
